Add files via upload
Obstacle Detector Using Arduino.
 
I have added 7 LEDs 1 LCD 7 1k Resistor and 1 active Buzzer if the obstacle is more than 25cm all the LEDs will be in OFF mode which means no obstacles detected within a radius of 25 Cm. If an obstacle is detected within 25cm it will show the distance at the LCD screen which is connected to Arduino using the I2c Protocol if the obstacle is less than 5 cm alarm will start beeping which alerts the user that an obstacle is there within a radius of 25Cm.

This project has various applications like it can be used as a reverse sensor in the car so that drivers can understand that an obstacle is there in a particular radius. It also can be used to find the underwater depth of oceans. It can also be implemented on ships to detect an iceberg or some potential threat ahead which can cause accidents.